+++ app.js
@@ -15,10 +15,8 @@ var path = require('path'),
     multer = require('multer') //사진 업로드
 
 var mysql = require('mysql');
-
 const mySqlClient = mysql.createConnection(require('./config/db_config'));
-const authConfig = require('./config/auth_config');
-const superPassword = authConfig.super_password;
+
 
 app.set('port', process.env.PORT || 3000);
 app.use(bodyParser.urlencoded({
@@ -36,13 +34,7 @@ app.use(cors());
 
 var router = express.Router();
 
-//MainPage 라우터
-router.route('/').get(function (req, res) {
 
-    fs.readFile('./public/index.html', 'utf8', function (error, data) {
-        res.send(ejs.render(data, {}));
-    });
-});
 
 //도면 이동 라우터
 router.route('/defact/drawing/').get(function (req, res) {
@@ -177,154 +169,28 @@ router.route('/defact/add/').get(function (req, res) {
     });
 });
 
-//회원가입 이동 라우터
-router.route('/register').get(function (req, res) {
-    fs.readFile('./public/register.html', 'utf8', function (error, data) {
-        res.send(ejs.render(data, {}));
-    });
-});
-
-
-//회원가입 라우터
-router.route('/reg_submit').post(function (req, res) {
-    var id = req.body.id;
-    var pw = req.body.pw;
-    var name = req.body.name;
-    var tel = req.body.tel;
-    var auth = req.body.auth;
-    var type = req.body.type;
-    var result;
-    var params = {
-        user_id: id,
-        password: pw,
-        name: name,
-        tel: tel,
-        type: type
-    };
-
-    var alertMsg = "";
-    if (auth == superPassword) {
-        result = checkInput(params);
-        if (result == 2) {
-            alertMsg = '연락처 입력이 잘못되었습니다.';
-            res.send('<script type="text/javascript">alert("' + alertMsg + '"); window.location="/register";</script>');
-        } else if (result == 3) {
-            alertMsg = '아이디는 영어/숫자 4-12자리로 입력해주세요.';
-            res.send('<script type="text/javascript">alert("' + alertMsg + '"); window.location="/register";</script>');
-        } else {
-            var checkIdSql = 'SELECT * FROM user WHERE user_id = ?;';
-            mySqlClient.query(checkIdSql, params.user_id, function (err, rows) {
-                if (err) {
-                    console.log("Search Error>>" + err);
-                    alertMsg = "회원등록 중 오류가 발생했습니다.";
-                    res.send('<script type="text/javascript">alert("' + alertMsg + '"); window.location="/register";</script>');
-                } else {
-                    if (rows.length > 0) {
-                        alertMsg = "이미 사용중인 아이디입니다.";
-                        res.send('<script type="text/javascript">alert("' + alertMsg + '"); window.location="/register";</script>');
-                    } else {
-                        var insertSql = 'INSERT INTO user SET ?;';
-                        mySqlClient.query(insertSql, params, function (err) {
-                            if (err) {
-                                console.log("Insert Error>>" + err);
-                                alertMsg = "회원등록 중 오류가 발생했습니다.";
-                                res.send('<script type="text/javascript">alert("' + alertMsg + '"); window.location="/register";</script>');
-                            } else {
-                                alertMsg = "회원가입이 완료되었습니다.";
-                                res.send('<script type="text/javascript">alert("' + alertMsg + '"); window.location="/";</script>');
-
-                            }
-                        });
-                    }
-
-                }
-            });
-        }
-    } else {
-        res.send('<script type="text/javascript">alert("인증코드가 일치하지 않습니다."); window.location="/register";</script>');
-        res.end();
-    }
-});
-
-//회원가입 이동 라우터
-router.route('/register').get(function (req, res) {
-    fs.readFile('./public/register.html', 'utf8', function (error, data) {
-        res.send(ejs.render(data, {}));
-    });
-});
-
-//회원가입 체크 라우터
-function checkInput(params) {
-    var result;
-    var pattern_eng = /[a-zA-Z]/; //영어 입력 체크 패턴
-    var pattern_num = /[0-9]/; //숫자 입력 체크 
-    var pattern_spc = /[~!@#$%^&*()_+|<>?:{}]/; // 특수문자
-    var pattern_kor = /[ㄱ-ㅎ|ㅏ-ㅣ|가-힣]/; // 한글체크
-    id = params.user_id;
-    tel = params.tel;
-    if ((pattern_num.test(id)) || (pattern_eng.test(id)) && !(pattern_spc.test(id)) && !(pattern_kor.test(id))) { //id검사
-        if ((pattern_num.test(tel)) && !(pattern_eng.test(tel)) && !(pattern_spc.test(tel)) && !(pattern_kor.test(tel))) { //tel검사
-            result = 1; //정상
-        } else {
-            result = 2; // tel 입력 오류
-        }
-    } else {
-        result = 3; //id 입력 오류
-    }
-    return result;
-}
-
-
-//-------------------------회원가입------------------------------------
-
-
 //--------------------------로그인-------------------------------------
-router.route('/process/login').post(function (req, res) {
-    if (req.session.user) {
-        console.log('세션 유저데이터 있음');
-        res.redirect('/selectc');
-    } else {
-        var checkId = req.body.id;
-        var checkPwd = req.body.password;
-        var selectPwdSql = "select * from user where user_id = ? && password=?";
-        mySqlClient.query(selectPwdSql, [checkId, checkPwd], function (err, row) {
-            if (err) {
-                console.log("dong/ho select page sql ERROR>>" + err);
-            } else {
-                if (row[0]) {
-                    req.session.user = {
-                        id: row[0].id,
-                        userId: checkId,
-                        userName: row[0].name,
-                        userType: row[0].type
-                    };
-
-                    res.redirect('/select');
-                    return true;
-                } else {
-                    res.send('<script type="text/javascript">alert("아이디 또는 비밀번호가 일치하지 않습니다."); window.location="/";</script>');
-                }
-            }
-        });
-    }
-});
+router.route('/process/login').post();
 
 //공사종류, 동, 호  선택페이지
 router.route('/select').get(function (req, res) {
     if (req.session.user) {
+        //동, 호 선택
         if (req.query.ctype) {
             fs.readFile('./public/select.html','utf8',function(err,data){
+                //선택한 공사종류를 쿠키에 저장
+               res.cookie('ctypeCookie',req.query.ctype);
                 res.send(ejs.render(data, {
-                    ctype: req.query.ctype,
                     name: req.session.user.userName,
                     type: req.session.user.userType
                 }));
             });
-        } else {
+        } 
+        //공사종류 선택
+        else {
             fs.readFile('./public/select_const.html', 'utf8', function (error, data) {
                 var ctype = req.query.ctype;
                 res.send(ejs.render(data, {
-                    ctype: ctype,
                     name: req.session.user.userName,
                     type: req.session.user.userType
                 }));
@@ -356,8 +222,17 @@ router.route('/defact/add_submit').post(function (req, res) {
 
 });
 
+var index = require('./routes/index.js');
+router.route('/').get(index);
+var register = require('./routes/register.js');
+router.route('/register').get(register.register);
+router.route('/reg_submit').post(register.reg_submit);
+var login = require('./routes/login.js');
+router.route('/process/login').post(login);
 
 app.use('/', router);
+
+
 
 // 404 에러 페이지 처리
 var errorHandler = expressErrorHandler({
